Manage floating points

+++ calculator/Program.cs
@@ -7,8 +7,8 @@ namespace calculator
         static void Main(string[] args)
         {
             bool again = false;
-            int number1 = 0;
-            int finalResult = 0;
+            double number1 = 0;
+            double finalResult = 0;
             Tools tool = new Tools();
             do
             {
@@ -21,8 +21,8 @@ namespace calculator
                 {
                     number1 = tool.AskUserForNumber("Enter the first number : ");
                 }
-                int userOperator = tool.AskUserForOperation();
-                int number2 = tool.AskUserForNumber("Enter the second number : ");
+                double userOperator = tool.AskUserForOperation();
+                double number2 = tool.AskUserForNumber("Enter the second number : ");
                 finalResult = tool.DisplayResult(userOperator, number1, number2);
                 Console.WriteLine("Your result : " + finalResult);
                 Console.WriteLine();
